# docs/openapi.yaml
openapi: 3.1.0
info:
  title: NyumbaLink API
  version: "1.0.0"
  description: |
    ## Kenya's Property Marketplace REST API

    **Base URL:** `https://nyumbalink.co.ke/api`

    ### Authentication
    All protected endpoints require a JWT Bearer token obtained from `/auth/login`.
    Include it as: `Authorization: Bearer <access_token>`

    ### Rate Limiting
    | Scope           | Limit             |
    |-----------------|-------------------|
    | Global          | 200 req/day, 50/hr |
    | Auth endpoints  | 10 req/min        |
    | File uploads    | 20 req/hr         |

    ### Pagination
    List endpoints accept:
    | Param     | Type    | Default | Description              |
    |-----------|---------|---------|--------------------------|
    | page      | integer | 1       | Page number (1-indexed)  |
    | per_page  | integer | 12      | Items per page (max 50)  |
    | sort_by   | string  | created_at | Column to sort by     |
    | sort_dir  | string  | desc    | `asc` or `desc`          |

  contact:
    name: NyumbaLink Support
    email: support@nyumbalink.co.ke
    url: https://nyumbalink.co.ke

  license:
    name: Proprietary
    url: https://nyumbalink.co.ke/terms
servers:
  - url: https://nyumbalink.co.ke/api
    description: Production
  - url: http://localhost:5000/api
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

# ──────────────────────────────────────────────────────────────────────
#  Reusable schemas
# ──────────────────────────────────────────────────────────────────────
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        total:     { type: integer, example: 248 }
        page:      { type: integer, example: 1 }
        per_page:  { type: integer, example: 12 }
        pages:     { type: integer, example: 21 }
        has_prev:  { type: boolean, example: false }
        has_next:  { type: boolean, example: true }
        sort_by:   { type: string,  example: created_at }
        sort_dir:  { type: string,  enum: [asc, desc], example: desc }
        prev_url:  { type: string,  nullable: true }
        next_url:  { type: string,  nullable: true }
...
    PropertyImage:
      type: object
      properties:
        id:         { type: string, format: uuid }
        image_url:  { type: string, format: uri }
        is_primary: { type: boolean }
        sort_order: { type: integer }
    Property:
      type: object
      properties:
        id:                  { type: string, format: uuid }
        title:               { type: string, example: "3 Bed Apartment – Kilimani" }
        description:         { type: string, nullable: true }
        category:            { type: string, enum: [rent, short_stay, plot_sale] }
        price:               { type: number, format: float, example: 45000 }
        price_period:        { type: string, enum: [monthly, per_night, yearly, total], nullable: true }
        location:            { type: string, example: "Kilimani, Nairobi" }
        county:              { type: string, example: "Nairobi" }
        town:                { type: string, example: "Kilimani" }
        latitude:            { type: number, nullable: true }
        longitude:           { type: number, nullable: true }
        bedrooms:            { type: integer, nullable: true }
        bathrooms:           { type: integer, nullable: true }
        plot_size:           { type: number, nullable: true }
        plot_size_unit:      { type: string, nullable: true }
        floor_area:          { type: number, nullable: true }
        amenities:           { type: array, items: { type: string } }
        owner_id:            { type: string, format: uuid }
        verification_status: { type: string, enum: [pending, verified, rejected] }
        status:              { type: string, enum: [active, inactive, sold, rented] }
        featured:            { type: boolean }
        views_count:         { type: integer }
        images:              { type: array, items: { $ref: '#/components/schemas/PropertyImage' } }
        primary_image:       { $ref: '#/components/schemas/PropertyImage', nullable: true }
        avg_rating:          { type: number, nullable: true }
        created_at:          { type: string, format: date-time }
        updated_at:          { type: string, format: date-time }
...
paths:
...
  /search:
    get:
      summary: Elasticsearch-powered advanced property search
      operationId: advancedSearch
      tags: [Search]
      parameters:
        - { name: q,          in: query, schema: { type: string }, description: Full-text search query }
        - { name: category,   in: query, schema: { type: string, enum: [rent, short_stay, plot_sale] } }
        - { name: county,     in: query, schema: { type: string } }
        - { name: min_price,  in: query, schema: { type: number } }
        - { name: max_price,  in: query, schema: { type: number } }
        - { name: bedrooms,   in: query, schema: { type: integer } }
        - { name: page,       in: query, schema: { type: integer, default: 1 } }
        - { name: per_page,   in: query, schema: { type: integer, default: 12 } }
      responses:
        "200":
          description: Search results with pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:       { type: array, items: { $ref: '#/components/schemas/Property' } }
                  pagination: { $ref: '#/components/schemas/PaginationMeta' }
                  query:      { type: string }
                  filters:    { type: object }

  # ── Inquiries ─────────────────────────────────────────────────────
...
  /notifications:
    get:
      summary: Get notification feed
      operationId: listNotifications
      tags: [Notifications]
      security:
        - BearerAuth: []
      parameters:
        - { name: page,     in: query, schema: { type: integer } }
        - { name: per_page, in: query, schema: { type: integer } }
      responses:
        "200":
          description: Notifications with unread count
          content:
            application/json:
              schema:
                type: object
                properties:
                  notifications: { type: array }
                  unread_count:  { type: integer }
                  pagination:    { $ref: '#/components/schemas/PaginationMeta' }
  /notifications/read-all:
    put:
      summary: Mark all notifications as read
      operationId: markAllRead
      tags: [Notifications]
      security:
        - BearerAuth: []
      responses:
        "200": { description: All marked as read }